Do not update if key is being held down
keyPressed() is called repeatedly if the key is held down. The
keypad control requires that the key will be held down for as
long as that value needs to be sent, but only needs to be sent once.

So, if the value is already set and sent, don't update it again.